put callback managing utils in seperate file
