+++ src/CMakeLists.txt
@@ -14,6 +14,7 @@ endif (NOT ANDROID)
 
 add_library(${PLUGIN_NAME} SHARED
   "wrapper.cc"
+  "wrapper_utils.cc"
   "../third_party/cronet_impl/sample_executor.cc"
   "${CMAKE_CURRENT_SOURCE_DIR}/../third_party/dart-sdk/dart_api_dl.c"
 )
